+++ app/commands/homebrew_context/daggerheart/copy_community_command.rb
@@ -2,15 +2,15 @@
 
 module HomebrewContext
   module Daggerheart
-    class CopyRaceCommand < BaseCommand
+    class CopyCommunityCommand < BaseCommand
       include Deps[
-        add_race: 'commands.homebrew_context.daggerheart.add_race',
+        add_community: 'commands.homebrew_context.daggerheart.add_community',
         add_feat: 'commands.homebrew_context.daggerheart.add_feat'
       ]
 
       use_contract do
         params do
-          required(:race).filled(type?: ::Daggerheart::Homebrew::Race)
+          required(:community).filled(type?: ::Daggerheart::Homebrew::Community)
           required(:user).filled(type?: ::User)
         end
       end
@@ -19,25 +19,25 @@ module HomebrewContext
 
       def do_persist(input)
         result = ActiveRecord::Base.transaction do
-          race = add_race.call({ user: input[:user], name: input[:race].name })[:result]
+          community = add_community.call({ user: input[:user], name: input[:community].name })[:result]
 
-          input[:race].user.feats.where(origin: 0, origin_value: input[:race].id).find_each do |feat|
-            add_feat.call(feat_attributes(feat, race).merge({ user: input[:user] }))
+          input[:community].user.feats.where(origin: 1, origin_value: input[:community].id).find_each do |feat|
+            add_feat.call(feat_attributes(feat, community).merge({ user: input[:user] }))
           end
 
-          race
+          community
         end
 
         { result: result }
       end
 
-      def feat_attributes(feat, race)
+      def feat_attributes(feat, community)
         feat
           .attributes
           .slice('origin', 'kind', 'limit', 'limit_refresh')
           .symbolize_keys
           .merge({
-            origin_value: race.id,
+            origin_value: community.id,
             title: feat.title['en'],
             description: feat.description['en'],
             limit: feat.description_eval_variables['limit']