+++ app/serializers/daggerheart/homebrew/community_serializer.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Daggerheart
+  module Homebrew
+    class CommunitySerializer < ApplicationSerializer
+      attributes :id, :name
+    end
+  end
+end