+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- managing homebrew communities for Daggerheart
+
 ## [0.3.0] - 2025-09-10
 
 ## [0.2.23] - 2025-09-08
@@ -16,7 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - homebrew sorting
 
 ### Fixed
-- rendering daggerheart PDF for characte with homebrew race
+- rendering daggerheart PDF for character with homebrew race
 
 ## [0.2.22] - 2025-09-05
 ### Fixed